move build_and_reuse_placeholder to input_source
