+++ docs/modules/graph_builder.rst
@@ -1,6 +1,9 @@
 tensorpack.graph_builder package
 ================================
 
+These are some useful functions if you need to write your own trainers.
+Note that they may not be well maintained.
+
 .. automodule:: tensorpack.graph_builder
     :members:
     :undoc-members:
