+++ tensorpack/input_source/input_source_base.py
@@ -12,9 +12,38 @@ from ..callbacks.base import CallbackFactory
 from ..tfutils.common import get_op_tensor_name
 from ..utils import logger
 from ..utils.argtools import call_only_once, memoized_method
-from ..graph_builder.model_desc import build_or_reuse_placeholder
+from ..compat import tfv1
 
 __all__ = ['InputSource', 'remap_input_source']
+
+
+def build_or_reuse_placeholder(tensor_spec):
+    """
+    Build a tf.placeholder from the metadata in the given tensor spec, or return an existing one.
+
+    Args:
+        tensor_spec (tf.TensorSpec):
+
+    Returns:
+        tf.Tensor:
+    """
+    g = tfv1.get_default_graph()
+    name = tensor_spec.name
+    try:
+        tensor = g.get_tensor_by_name(name + ':0')
+        assert "Placeholder" in tensor.op.type, "Tensor {} exists but is not a placeholder!".format(name)
+        assert tensor_spec.is_compatible_with(tensor), \
+            "Tensor {} exists but is not compatible with the signature!".format(tensor)
+        if tensor.shape == tensor_spec.shape:
+            # It might be desirable to use a placeholder of a different shape in some tower
+            # (e.g., a less specific shape)
+            return tensor
+    except KeyError:
+        pass
+    with tfv1.name_scope(None):   # clear any name scope it might get called in
+        ret = tfv1.placeholder(
+            tensor_spec.dtype, shape=tensor_spec.shape, name=tensor_spec.name)
+    return ret
 
 
 def get_tensors_inputs(placeholders, tensors, names):
